DICT
Hi guys, i'm new to programming and to python. This is an Assignment on a course im doing. I have an error on the 2nd for loop, the nested one. It happens that i have to get the email addresses, those which are the 2nd word of each line starting with "From ". The problem is that when i call to add just the parse[1] word, which are the addresses, it adds all to the dict(counts).

The output isn't the desired, but it isn't so far neither, it's just that it takes the whole strings in the splitted lines, and ends up showing "From 27" instead of "email-adress count", which is the key/value i need to get.

Thx in advance.